Update patch set 2
Patch Set 2:

> (1 comment)

We have the same issue today: if someone modifies the Gopkg.toml file 'make populate' will ignore the changes as the make target runs 'dep ensure --vendor-only' which converts the .lock file into a vendor directory.

The user is expected to run 'dep ensure' after changing a constraint in the toml file to sync the lock file and update the vendor directory with the new constraints. All of this behaviour is retained by this changeset.

Patch-set: 2
Label: Code-Review=0