[DEV-9337] Re-connect event client to local peer
Use the SDK's 'prefer-peer' resolver. This peer resolver
initially connects to the local peer if the local peer's
block height is within a certain threshold of the block
height of the channel. If the local peer's block height falls
below this threshold then the event client disconnects from
the local peer and attempts to connect to another peer within
the local org. Once the local peer's block height has caught up,
the event client reconnects to the local peer.

Change-Id: If3fd91661cb52a8ebbde130f64bf61449d10d5d4
Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>